Fix rho and added first version of midppl-tree-parer.
